@@ -1,7 +1,7 @@
-import { inject } from '@angular/core';
 import { CanActivateFn } from '@angular/router';
 import { FirebaseService } from '../services/firebase.service';
 import { UtilsService } from '../services/utils.service';
+import { inject } from '@angular/core';
 
 export const authGuard: CanActivateFn = async (route, state) => {
   const firebaseService = inject(FirebaseService);
