@@ -2,5 +2,6 @@ export interface User {
   uid: string,
   email: string,
   password: string,
-  name: string
+  name: string,
+  image: string
 }
