@@ -8,6 +8,7 @@ import {
   IonBackButton,
   IonButton,
   IonIcon,
+  IonMenuButton
 } from '@ionic/angular/standalone';
 import { addIcons } from 'ionicons';
 import { closeCircleOutline } from 'ionicons/icons';
@@ -19,20 +20,22 @@ import { UtilsService } from 'src/app/services/utils.service';
   styleUrls: ['./header.component.scss'],
   standalone: true,
   imports: [
+    IonIcon,
+    IonButton,
     IonBackButton,
     IonButtons,
     IonTitle,
     IonToolbar,
     IonHeader,
     CommonModule,
-    IonIcon,
-    IonButton,
+    IonMenuButton
   ],
 })
 export class HeaderComponent implements OnInit {
   @Input({ required: true }) title!: string;
   @Input() backButtonURL: string | null = null;
   @Input() isModal: boolean = false;
+  @Input() showMenuButton: boolean = false;
 
   utilsService = inject(UtilsService);
 
@@ -40,9 +43,7 @@ export class HeaderComponent implements OnInit {
     addIcons({ closeCircleOutline });
   }
 
-  ngOnInit() {
-    console.log("Hola")
-  }
+  ngOnInit() {}
 
   dismissModal() {
     this.utilsService.dismissModal();
